server/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
psycopg2-binary
torchcodec
supabase
pyjwt
bcrypt
boto3
email-validator
python-multipart
requests
langchain-google-genai
google-genai
reportlab
reportlab